requirements.txt
Django==4.2.7
pymysql==1.1.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7

